Shutdown server when port has been changed
Since a1e9ed26c31d375cc0e454084c0e7c7f64cd20f7 the port of the SFTP
server can be changed during test. In this case the SFTP server was not
shutdown after the test, because the rule restarted it but Apache SSHD
is not designed to be restarted. See
https://issues.apache.org/jira/browse/SSHD-769

We now start a new server when the port is changed. This server is
shutdown after the test.